refactor(knowledge): rename jobs dir to tasks per handler-authoring convention
Aligns the knowledge job-handler directory with the canonical
`<module>/tasks/<Name>JobHandler.ts` layout documented in
docs/references/job-and-scheduler/handler-authoring.md §6.
