+++ src/main/features/knowledge/tasks/__tests__/deleteSubtreeJobHandler.test.ts
