+++ src/main/features/knowledge/tasks/__tests__/indexDocumentsJobHandler.test.ts
