+++ src/main/features/knowledge/tasks/__tests__/reindexSubtreeJobHandler.test.ts
