+++ src/main/features/knowledge/tasks/deleteSubtreeJobHandler.ts
