+++ src/main/features/knowledge/tasks/indexDocumentsJobHandler.ts
