+++ src/main/features/knowledge/tasks/prepareRootJobHandler.ts
