+++ src/main/features/knowledge/tasks/reindexSubtreeJobHandler.ts
