+++ src/main/features/knowledge/tasks/utils/cancel.ts
