+++ src/main/features/knowledge/tasks/utils/jobInput.ts
