+++ src/main/features/knowledge/tasks/utils/settled.ts
